pyproject.toml
[tool.poetry]
name = "gojira"
version = "1.0.0"
description = "A anime bot for Telegram made in Python"
authors = ["Hitalo"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "4.9.3"
httpx = {version = "0.22.0", extras = ["http2"]}
humanize = "3.11.0"
meval = "2.5"
PyYAML = "5.4.1"
psutil = "5.8.0"
TgCrypto = "1.2.3"
telegraph = "1.4.1"
python-anilist = "1.0.9"
uvloop = {version = "^0.16", optional = true}
pyrogram = {git = "https://github.com/AmanoTeam/Pyrogram"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"

[tool.poetry.extras]
fast = ["uvloop"]

[tool.black]
exclude = '''
/(
    \.git
  | \.pytest_cache
  | \.vscode
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
